added subplot convenience functions
Summary of changes:
1. Used matplotlib.gridpec to create a gridspec object that keeps track
of the desired layout. Can call a layout creation by calling the method
make_subplots with arguments passable to gridspec.

2. moved set_font call in add_plot to be above the axes function calls
as with subplots, the font size was not correctly set if the call was
made after making the axes function calls.

3. added a set_subplot method to allow user to select the subplot axis
on which to add_plot

4. added example for usage in the subplots section of the jupyter
notebook

5. added a line plt.ion to automatically turn on interactive mode if it
is not already on as even the basic plotting was not working without
this mode on